getting results from API -> loading into Card component
