+++ src/App.jsx
@@ -3,6 +3,8 @@ import SignInPage from "./pages/auth/SigninPage";
 import SignUpPage from "./pages/auth/SignUpPage";
 import Dashboard from "./pages/Dashboard";
 import { PortfolioProvider } from "./contexts/PortfolioContext";
+import ProductSearch from "./pages/Product-Search-Page/ProductSearch";
+
 
 function App() {
   return (
@@ -13,6 +15,7 @@ function App() {
           <Route path="/" element={<SignInPage />} />
           <Route path="/sign-up" element={<SignUpPage  />} />
           <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/search" element={<ProductSearch />} />
         </Routes>
       </BrowserRouter>
      </PortfolioProvider>
